refactor: move time-dependent study selection to entities
