+++ src/entities/deck/index.ts
@@ -6,9 +6,8 @@ export {
   getCategory,
   isHighlightLanguage,
   mustFindDeckById,
-  selectStudyCardsForDeck,
 } from "./model/rules";
-export { useDeck, useDecks } from "./model/hooks";
+export { useDeck, useDecks, useFilteredStudyCards } from "./model/hooks";
 export { deckFormSchema } from "./model/schema";
 export { clearRemoteDecks } from "./model/store";
 export type { Deck, DeckCreateInput, DeckEdit, DeckId, LocalDeckCreateInput, RemoteDeck } from "./model/types";
