+++ src/entities/deck/model/hooks.ts
@@ -1,5 +1,9 @@
 import { useStore } from "zustand";
 
+import type { Card } from "@/entities/card/@x/deck";
+import { useTimeDependentValue } from "@/shared/lib/useTimeDependentValue";
+
+import { selectStudyCardsForDeck } from "./rules";
 import { deckStore } from "./store";
 import type { Deck, DeckId } from "./types";
 
@@ -13,3 +17,13 @@ export const useDeck = (id: DeckId | undefined): Deck | undefined =>
     deckStore,
     (state) => state.remoteDecks.find((deck) => deck.id === id) ?? state.localDecks.find((deck) => deck.id === id)
   );
+
+export const useFilteredStudyCards = (
+  deck: Deck | undefined,
+  cards: Card[],
+  preferences: { study: { useCardInterval: boolean } }
+): Card[] =>
+  useTimeDependentValue((now) => {
+    const selection = selectStudyCardsForDeck(cards, deck, preferences.study, now);
+    return { value: selection.cards, nextUpdateAt: selection.nextAvailabilityAt };
+  });
