+++ src/features/deck-filter/index.ts
@@ -1,3 +1,2 @@
 export { useDeckFilterState } from "./model/useDeckFilterState";
-export { useFilteredStudyCards } from "./model/useFilteredStudyCards";
 export { DeckFilterForm } from "./ui/DeckFilterForm";
